Move UWP code to shared project
