+++ uwp/ReactiveUI.UwpRouting.Shared/App.xaml
@@ -1,5 +1,8 @@
 ﻿<Application
     x:Class="ReactiveUI.UwpRouting.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ReactiveUI.UwpRouting"
+    RequestedTheme="Light">
+
 </Application>
