+++ uwp/ReactiveUI.UwpRouting.Shared/Assets/lockscreenlogo.scale-200.png
