+++ uwp/ReactiveUI.UwpRouting.Shared/Assets/splashscreen.scale-200.png
