+++ uwp/ReactiveUI.UwpRouting.Shared/Assets/square150x150logo.scale-200.png
