+++ uwp/ReactiveUI.UwpRouting.Shared/Assets/storelogo.png
