+++ uwp/ReactiveUI.UwpRouting.Shared/Assets/wide310x150logo.scale-200.png
