+++ uwp/ReactiveUI.UwpRouting.Shared/ViewModels/MainViewModel.cs
