+++ uwp/ReactiveUI.UwpRouting.UWP/Package.appxmanifest
