+++ uwp/ReactiveUI.UwpRouting.UWP/Properties/AssemblyInfo.cs
