--- uwp/ReactiveUI.UwpRouting/App.xaml.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using Windows.ApplicationModel.Activation;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Navigation;
-using ReactiveUI.UwpRouting.Views;
-
-namespace ReactiveUI.UwpRouting
-{
-    public sealed partial class App : Application
-    {
-        public App() => InitializeComponent();
-
-        protected override void OnLaunched(LaunchActivatedEventArgs e)
-        {
-            if (!(Window.Current.Content is Frame rootFrame))
-            {
-                rootFrame = new Frame();
-                rootFrame.NavigationFailed += OnNavigationFailed;
-                Window.Current.Content = rootFrame;
-            }
-
-            if (e.PrelaunchActivated != false) return;
-            if (rootFrame.Content == null) rootFrame.Navigate(typeof(MainView), e.Arguments);
-            Window.Current.Activate();
-        }
-
-        private static void OnNavigationFailed(object sender, NavigationFailedEventArgs e)
-        {
-            throw new Exception("Failed to load Page " + e.SourcePageType.FullName);
-        }
-    }
-}
